isize -> f64 pour les nombres (type définitif)

+++ app-poc-rust/src/communs.rs
@@ -1,7 +1,7 @@
 
 #[derive(Debug)]
 pub enum Types {
-	Nombre(isize), 
+	Nombre(f64), 
 }
